Various changes to CYRC01, made timers better
The second jump pad in the RJVoid section of CYRC01 now has a wall in front of
it, to help direct attention to your right. Aerial and turret drones have been
added to the RJVoid section, as well.

The timer displays have been made more accurate, but the timers themselves are
less accurate, so that they can exceed 15 minutes without overflowing.

Also, a Log wasn't cyber_cl_debug'd. Fixed.

Fixes #19.