Added cyber_mph_barscale and cyber_mph_friction
cyber_mph_barscale is a control for how far the bar goes per unit/tic, equal to how many units/tic you need for the bar to wrap around.

cyber_mph_friction controls how resistant the bar is to changes, and it is a percentage. Default is 300, meaning it moves 1/3 of the way to your true speed every tic, with an auto-snap when it gets close enough.