Added a secret area to CYRC01 + a way back to RJSpace + more goodies at the end
The secret area contains the Instagib Carbine, an instagib rifle with only 20 shots and a slower rate of fire.
It is unaffected by full ammo pickups, so if you want to refill it, you have to go all the way back to the instagib pickup.
It's also a normal weapon pickup for now.